modelshield-api/requirements.txt
fastapi==0.103.2
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic==1.10.14
bcrypt==4.0.1